+++ lib/src/context_log.c
@@ -67,7 +67,7 @@ void clContextDefaultLog(clContext * C, const char * section, int indent, const 
 {
     if (section) {
         char spaces[10] = "         ";
-        int spacesNeeded = 9 - strlen(section);
+        int spacesNeeded = 9 - (int)strlen(section);
         spacesNeeded = CL_CLAMP(spacesNeeded, 0, 9);
         spaces[spacesNeeded] = 0;
         fprintf(stdout, "[%s%s] ", spaces, section);
